Modularize utils
